requirements.txt
streamlit==1.2.0  # Adjust this to the version you are using
sqlite3==3.36.0  # This is part of the Python standard library, but some platforms might require it
pandas==1.3.3
google-cloud-storage==1.42.3
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-core==2.2.1
google-cloud-core==2.1.0
